backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
scikit-learn
numpy
pytest
httpx
